refactor: Extract sub-components from command-palette
